refactor: dialog container shared component화 및 적용
