--- src/shared/ui/dialog/dialog.ui.ts
@@ -1,3 +1,0 @@
-import * as DialogPrimitive from "@radix-ui/react-dialog"
-
-export const Dialog = DialogPrimitive.Root
